backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
requests==2.31.0
python-dotenv==1.0.0
python-multipart==0.0.6
Pillow==10.2.0
SQLAlchemy==2.0.25
pymysql==1.1.0
cryptography==42.0.2
PyJWT==2.8.0
google-auth==2.27.0
